# Conflicts:
#	app/app.iml
#	app/build.gradle
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/easyfitness/CardioCursorAdapter.java
#	app/src/main/java/com/easyfitness/CountdownDialogbox.java
#	app/src/main/java/com/easyfitness/DAO/DAOFonte.java
#	app/src/main/java/com/easyfitness/DAO/DAOProfil.java
#	app/src/main/java/com/easyfitness/DAO/DatabaseHelper.java
#	app/src/main/java/com/easyfitness/DAO/Fonte.java
#	app/src/main/java/com/easyfitness/DAO/Profile.java
#	app/src/main/java/com/easyfitness/MainActivity.java
#	app/src/main/java/com/easyfitness/ProfilFragment.java
#	app/src/main/java/com/easyfitness/ProfilWeightCursorAdapter.java
#	app/src/main/java/com/easyfitness/fonte/FonteCursorAdapter.java
#	app/src/main/java/com/easyfitness/fonte/FonteGraphFragment.java
#	app/src/main/java/com/easyfitness/fonte/FontesFragment.java
#	app/src/main/java/com/easyfitness/graph/CustomMarkerView.java
#	app/src/main/java/com/easyfitness/graph/Graph.java
#	app/src/main/java/com/easyfitness/machines/MachineDetailsFragment.java
#	app/src/main/java/com/easyfitness/utils/MusicController.java
#	app/src/main/res/drawable-mdpi/ic_star_black_24dp.png
#	app/src/main/res/drawable-xhdpi/marker2.png
#	app/src/main/res/layout/activity_main.xml
#	app/src/main/res/layout/dialog_rest.xml
#	app/src/main/res/layout/graph_markerview.xml
#	app/src/main/res/layout/machine_details.xml
#	app/src/main/res/layout/tab_fontes.xml
#	app/src/main/res/layout/tab_graph.xml
#	app/src/main/res/values-fr/strings.xml
#	app/src/main/res/values/about.xml
#	app/src/main/res/values/strings.xml
#	app/src/main/res/xml/settings.xml
#	build/intermediates/dex-cache/cache.xml

+++ app/src/main/java/com/easyfitness/DAO/Fonte.java
@@ -10,7 +10,7 @@ public class Fonte {
     private String mMachine;
     private int mSerie;
     private int mRepetition;
-    private int mPoids;
+    private float mPoids;
     private int mUnit;
     private String mNote;
     private Profile mProfile;
@@ -20,7 +20,7 @@ public class Fonte {
     /*
        * Fonte(Date pDate, String pMachine, int pSerie, int pRepetition, int pPoids, Profile pProfile)
        */
-    public Fonte(Date pDate, String pMachine, int pSerie, int pRepetition, int pPoids, Profile pProfile, int pUnit, String pNote, long pMachineKey, String pTime) {
+    public Fonte(Date pDate, String pMachine, int pSerie, int pRepetition, float pPoids, Profile pProfile, int pUnit, String pNote, long pMachineKey, String pTime) {
         super();
         this.mDate = pDate;
         this.mMachine = pMachine;
@@ -62,7 +62,7 @@ public class Fonte {
         return mRepetition;
     }
 
-    public int getPoids() {
+    public float getPoids() {
         return mPoids;
     }
